Only bump state once per collection-type-change, at the end of a method call
If setCakes internally calls setCakeToppers once per cake, at the end we
should only bump cakeTopperState once for the entire collection.

If the internal calls fail (or the cookies themselves fail to make) after
creating the cake toppers, those toppers should be discarded and the
state should not change.

Also wrap ix_gets in these transactions as well - in the future we will
likely set the gets to SERIALIZABLE isolation level so that they get
repeatable reads, but for now this ensures that $ctx->state doesn't
hang around between separate method calls in a JMAP request