Rename methods to match JMAP names
I'm not sure why we decided not to do this when we did the big renaming.
This commit changes the method names (and their hooks) to better match
the JMAP verbs to which they correspond. That entails this set of
changes:

- `get_updates` is now `changes` (and related, so `ix_get_updates_check`
  is now `ix_changes_check`)

- `get_list_updates` is now `query_changes` (and so,
  `ix_query_changes_check`)

- `get_list` is now `query`. This was the biggest change, only because
  there are a bunch of auxiliary methods used for querying. These
  methods are now called ix_query_enabled, ix_query_check,
  ix_query_filter_map, ix_query_sort_map, and ix_query_joins.

This also changes the name of the tests so that they still make sense:
updates.t is now changes.t, and get_foo_list.t is now query.t.